[#14] feat : redis user data fetch 기능 완성
