+++ .env.development
@@ -1,0 +1,2 @@
+# 개발 환경 api url
+REACT_APP_API_URL=http://localhost:8080/api/
