+++ package.json
@@ -14,22 +14,28 @@
     "@mui/styled-engine-sc": "^6.0.2",
     "@mui/x-data-grid": "^7.18.0",
     "@mui/x-data-grid-generator": "^7.18.0",
-    "@mui/x-date-pickers": "^7.18.0",
+    "@mui/x-date-pickers": "^7.21.0",
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "@types/dompurify": "^3.0.5",
     "axios": "^1.7.7",
     "cors": "^2.8.5",
+    "date-fns": "^2.29.3",
+    "dompurify": "^3.1.7",
     "dotenv": "^16.4.5",
     "express": "^4.21.1",
+    "highlight.js": "^11.10.0",
     "http": "^0.0.1-security",
     "js-cookie": "^3.0.5",
     "jwt-decode": "^4.0.0",
     "mui-tel-input": "^6.0.1",
+    "quill": "^2.0.2",
     "react": "^18.3.1",
     "react-big-calendar": "^1.14.1",
     "react-dom": "^18.3.1",
     "react-grid-layout": "^1.4.4",
+    "react-quill": "^2.0.0",
     "react-router-dom": "^6.26.2",
     "react-scripts": "5.0.1",
     "react-syntax-highlighter": "^15.6.1",
@@ -64,6 +70,7 @@
   },
   "devDependencies": {
     "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
     "@eslint/js": "^9.10.0",
     "eslint": "^8.57.0",
     "eslint-config-prettier": "^9.1.0",
