+++ src/pages/login/Login.js
@@ -27,6 +27,8 @@ function Login() {
       const token = localStorage.getItem("token");
       const refreshToken = Cookies.get("refreshToken");
 
+      socket.emit("fetchData", token);
+
       console.log(token);
 
       if (token && isTokenValid(token)) {
